EDITED kbranch 50% done

+++ DentEngine/src/polyclipping/clipper.hpp
@@ -108,10 +108,10 @@ struct IntPoint {
   {
       return IntPoint(a.X-b.X, a.Y-b.Y);
   }
-  friend inline IntPoint operator/ (const IntPoint& a, const int d){
+  friend inline IntPoint operator/ (const IntPoint& a, const float d){
       return IntPoint(a.X/d, a.Y/d);
   }
-  friend inline IntPoint operator* (const IntPoint& a, const int d){
+  friend inline IntPoint operator* (const IntPoint& a, const float d){
       return IntPoint(a.X*d, a.Y*d);
   }
   friend inline IntPoint operator+= (IntPoint& a, const IntPoint& b){
